aiff_source: Improve robustness by allowing for SSND chunk to be at an arbitrary position
So far we assumed that the SSND chunk of an AIFF or AIFF-C file would be the last and we would skip any chunks coming after it. This would block us from loading files where the mandatory COMM chunk is after the SSND chunk, as we would miss crucial information like the overall number of frames or the sample rate. This change adds support for such files by looking for the SSND chunk only after reading all other supported chunks

Context: LIBS-898