Fix potential integer overflow at end of stream
If in SyncronousByteStream::Read, m_source.read() reaches the end of
stream, boost::iostreams::file_descriptor returns -1 as an end of stream
indicator. Previously, by casting the returned value to ULONG, an
integer overflow would occur. To fix this, the end of stream indicator
(-1) is caught and handled as "0 bytes read".